- using Feature2D everywhere to unify the interface ...
	modified:   ../cvt/CMakeLists.txt
	modified:   ../cvt/vision/AGAST.cpp
	modified:   ../cvt/vision/AGAST.h
	modified:   ../cvt/vision/Brief.cpp
	modified:   ../cvt/vision/Brief.h
	modified:   ../cvt/vision/FAST.cpp
	modified:   ../cvt/vision/FAST.h
	deleted:    ../cvt/vision/FeatureExtractor.h
	modified:   ../cvt/vision/Ferns.cpp
	modified:   ../cvt/vision/Ferns.h
	modified:   ../cvt/vision/PatchGenerator.cpp
	modified:   ../cvt/vision/PatchGenerator.h
	modified:   ../cvt/vision/agast/agast5_8.cpp
	modified:   ../cvt/vision/agast/agast7_12d.cpp
	modified:   ../cvt/vision/agast/agast7_12s.cpp
	modified:   ../cvt/vision/agast/oast9_16.cpp
	modified:   ../cvt/vision/fast/fast10.cpp
	modified:   ../cvt/vision/fast/fast11.cpp
	modified:   ../cvt/vision/fast/fast12.cpp
	modified:   ../cvt/vision/fast/fast9.cpp
	modified:   ../cvt/vision/fast/fast9simd.cpp
	modified:   ../tests/brief/brief.cpp
	modified:   ../tests/features/FeatureTracker.cpp
	modified:   ../tests/features/FeatureTracker.h
	modified:   ../tests/features/FeatureView.h
	modified:   ../tests/features/features.cpp
	modified:   ../tests/ferns/ferns.cpp

+++ cvt/vision/Brief.cpp
@@ -3,7 +3,7 @@
 #include <cvt/math/Math.h>
 
 #include <cvt/util/CVTTest.h>
-#include <cvt/io/Resources.h>   
+#include <cvt/io/Resources.h>
 
 namespace cvt {
     Brief::Brief( size_t patchSize ) : _patchSize( patchSize )
@@ -11,41 +11,41 @@ namespace cvt {
         //  make it odd!
         _patchSize |= 1;
         createTests();
-        
+
         _blurred.reallocate( _patchSize, _patchSize, IFormat::GRAY_UINT8 );
-        _tmp.reallocate( _patchSize, _patchSize, IFormat::GRAY_UINT8 );        
+        _tmp.reallocate( _patchSize, _patchSize, IFormat::GRAY_UINT8 );
     }
-    
+
     Brief::~Brief()
     {
     }
-    
+
     void Brief::descriptorForPatch( BriefDescriptor & d, const Image & patch )
     {
         // blur the patch
         patch.convolve( _tmp, IKernel::GAUSS_HORIZONTAL_7 );
         _tmp.convolve( _blurred, IKernel::GAUSS_VERTICAL_7 );
-        
+
         // perform the tests
         uint8_t signature = 0;
-        
+
         uint8_t * t0 = _tests;
         uint8_t * t1 = _tests + 2;
-        
+
         size_t stride;
         uint8_t * p = _blurred.map( &stride );
-        
+
         size_t i = 256;
         uint8_t descPart = 0;
         uint8_t descIter = 0;
-        
+
         while ( i-- ) {
             if( p[ t0[ 1 ] * stride + t0[ 0 ] ] > p[ t1[ 1 ] * stride + t1[ 0 ] ] )
                 signature |= 1;
             signature <<= 1;
             t0 += 4;
             t1 += 4;
-            
+
             descIter++;
             if( descIter == 8 ){
                 descIter = 0;
@@ -54,46 +54,46 @@ namespace cvt {
                 signature = 0;
             }
         }
-        
-        _blurred.unmap( p );        
+
+        _blurred.unmap( p );
     }
-    
-    
-    void Brief::descriptorForPatch( BriefDescriptor & d, const IntegralImage & image, const Vector2i & pos )
+
+
+    void Brief::descriptorForPatch( BriefDescriptor & d, const IntegralImage & image, const Vector2f & pos )
     {
         size_t iStride;
-        
+
         const float * ii = image.sumImage().map<float>( &iStride );
         size_t offset = -( _patchSize >> 1 ) - ( _patchSize >> 1 ) * iStride; // offset from patch center to upper right corner
-        
-        const float * p = ii + pos.y * iStride + pos.x + offset;
+
+        const float * p = ii + (int)pos.y * iStride + (int)pos.x + offset;
         const float * p0 = p;
         const float * p1 = p;
-        
+
         uint8_t * t0 = _tests;
         uint8_t * t1 = _tests + 2;
-        
+
         size_t i = 256;
         uint8_t descPart = 0;
         uint8_t descIter = 0;
-                
+
         uint64_t signature = 0;
         size_t subWindowSize = 1;
         while ( i-- ) {
             p0 = p + t0[ 1 ] * iStride + t0[ 0 ];
             p1 = p + t1[ 1 ] * iStride + t1[ 0 ];
-            
-            /*std::cout << "FEATURE: " << pos.x << ", " << pos.y << std::endl; 
-            std::cout << "TO: " << ( int )t0[ 0 ] << ", " << ( int )t0[ 1 ] << std::endl; 
+
+            /*std::cout << "FEATURE: " << pos.x << ", " << pos.y << std::endl;
+            std::cout << "TO: " << ( int )t0[ 0 ] << ", " << ( int )t0[ 1 ] << std::endl;
             std::cout << "T1: " << ( int )t1[ 0 ] << ", " << ( int )t1[ 1 ] << std::endl;
             std::cout << iStride << std::endl;*/
             if( IntegralImage::area( p0, subWindowSize, subWindowSize, iStride ) > IntegralImage::area( p1, subWindowSize, subWindowSize, iStride ) )
                 signature |= 1;
             signature <<= 1;
-            
+
             t0 += 4;
             t1 += 4;
-            
+
             descIter++;
             if( descIter == 64 ){
                 descIter = 0;
@@ -102,57 +102,57 @@ namespace cvt {
                 signature = 0;
             }
         }
-        
+
         image.sumImage().unmap( ii );
     }
-    
+
     void Brief::createTests()
     {
         RNG rng( time( NULL ) );
         double minmax = _patchSize >> 1;
         double sigma = Math::sqr( _patchSize ) / 25.0f;
-        
-        double x, y;     
+
+        double x, y;
         for( size_t i = 0; i < 256; i++ ){
             x = Math::clamp( rng.gaussian( sigma ), -minmax, minmax) + minmax;
-            y = Math::clamp( rng.gaussian( sigma ), -minmax, minmax) + minmax;            
+            y = Math::clamp( rng.gaussian( sigma ), -minmax, minmax) + minmax;
             _tests[ 4 * i ] = ( uint8_t )x;
             _tests[ 4 * i + 1 ] = ( uint8_t )y;
 
             x = Math::clamp( rng.gaussian( sigma ), -minmax, minmax) + minmax;
-            y = Math::clamp( rng.gaussian( sigma ), -minmax, minmax) + minmax;            
+            y = Math::clamp( rng.gaussian( sigma ), -minmax, minmax) + minmax;
             _tests[ 4 * i + 2 ] = ( uint8_t )x;
             _tests[ 4 * i + 3 ] = ( uint8_t )y;
         }
     }
-         
+
     BEGIN_CVTTEST( brief )
 
-    Brief b( 7 );    
+    Brief b( 7 );
     Resources resources;
     Image lena( resources.find( "lena_g.png" ) );
     Recti roi( 200, 100, 7, 7 );
     Image patch( lena, &roi );
-    
+
     BriefDescriptor d;
     b.descriptorForPatch( d, patch );
-    
+
     const uint8_t * desc = d.desc();
-   
-    BriefDescriptor d2( d );    
-    
-    
+
+    BriefDescriptor d2( d );
+
+
     bool ret = true;
-    ret &= ( d2.distance( d ) == 0 );    
+    ret &= ( d2.distance( d ) == 0 );
     uint64_t val = desc[ 0 ];
     val ^= 7;
     d2.set( val, 0 );
     ret &= ( d2.distance( d ) == 3 );
-    
+
     CVTTEST_PRINT( "Hamming Distance", ret );
-            
+
     return ret;
-    
+
     END_CVTTEST
 }
